fix: update imports after chunk_delta_h → chunk_delta_h_sm100 rename
